[dev] PSR-4 naming fixes in tests (part 2) (#51198)
In this PR we address PSR-4 naming-related violations (class FQN vs path mismatches and co) in tests.
