+++ plugins/woocommerce/changelog/dev-psr4-naming-fixes-part-2
@@ -1,0 +1,4 @@
+Significance: patch
+Type: dev
+
+Coding standatds: fix PSR-4 naming violation in tests (part 2).
