+++ plugins/woocommerce/tests/php/src/Blocks/MigrationTest.php
@@ -1,25 +1,47 @@
 <?php
+declare( strict_types = 1 );
 
-namespace Automattic\WooCommerce\Tests\Blocks\Utils;
+namespace Automattic\WooCommerce\Tests\Blocks;
 
 use Automattic\WooCommerce\Blocks\Migration;
 use Automattic\WooCommerce\Blocks\Options;
 use Mockery;
 
+/**
+ * Unit tests for the Migration class.
+ */
 class MigrationTest extends \WP_UnitTestCase {
 	use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 
+	/**
+	 * Setup testing environment.
+	 *
+	 * @return void
+	 */
 	protected function setUp(): void {
 		parent::setUp();
 		delete_option( Options::WC_BLOCK_USE_BLOCKIFIED_PRODUCT_GRID_BLOCK_AS_TEMPLATE );
 		delete_option( Options::WC_BLOCK_VERSION );
 	}
 
+	/**
+	 * Cleanup testing environment.
+	 *
+	 * @return void
+	 */
 	protected function tearDown(): void {
 		\Mockery::close();
 		parent::tearDown();
 	}
 
+	/**
+	 * Inject test data into migration mock.
+	 *
+	 * @param Migration $mock The mock to patch.
+	 * @param array     $data The data to inject.
+	 *
+	 * @return Migration
+	 */
 	private function set_db_upgrades( Migration $mock, array $data ) {
 		$reflection = new \ReflectionClass( Migration::class );
 		$property   = $reflection->getProperty( 'db_upgrades' );
@@ -32,6 +54,9 @@ class MigrationTest extends \WP_UnitTestCase {
 		return $mock;
 	}
 
+	/**
+	 * Test running single migration.
+	 */
 	public function test_migrations_run() {
 		update_option( Options::WC_BLOCK_VERSION, '1.0.0' );
 
@@ -53,6 +78,9 @@ class MigrationTest extends \WP_UnitTestCase {
 		$mock->run_migrations();
 	}
 
+	/**
+	 * Test running multiple migrations.
+	 */
 	public function test_multiple_migrations_run() {
 		update_option( Options::WC_BLOCK_VERSION, '0.0.9' );
 
@@ -76,6 +104,9 @@ class MigrationTest extends \WP_UnitTestCase {
 		$mock->run_migrations();
 	}
 
+	/**
+	 * Test skipping executed migrations.
+	 */
 	public function test_skip_migrations() {
 		update_option( Options::WC_BLOCK_VERSION, '2.0.0' );
 
@@ -95,6 +126,9 @@ class MigrationTest extends \WP_UnitTestCase {
 		$mock->run_migrations();
 	}
 
+	/**
+	 * Test skipping executed migrations.
+	 */
 	public function test_skip_migrations_when_missing_version_option() {
 
 		$mock = Mockery::mock( Migration::class )->makePartial();
@@ -112,5 +146,4 @@ class MigrationTest extends \WP_UnitTestCase {
 
 		$mock->run_migrations();
 	}
-
 }
