+++ plugins/woocommerce/tests/php/src/Blocks/Mocks/BlockHooksTestBlock.php
