+++ plugins/woocommerce/tests/php/src/Blocks/Mocks/OrderConfirmation/TotalsMock.php
@@ -1,4 +1,6 @@
 <?php
+declare( strict_types = 1 );
+
 namespace Automattic\WooCommerce\Tests\Blocks\Mocks\OrderConfirmation;
 
 use Automattic\WooCommerce\Blocks\Package;
@@ -23,15 +25,24 @@ class TotalsMock extends Totals {
 		);
 	}
 
-	public function render_content( $order, $permission = false, $attributes = [], $content = '' ) {
-		return parent::render_content( $order, $permission, $attributes, $content );
-	}
-
 	/**
 	 * For now don't need to initialize anything in tests so let's
 	 * just override the default behaviour.
 	 */
 	protected function initialize() {
 	}
-}
 
+	/**
+	 * This renders the content of the block within the wrapper.
+	 *
+	 * @param \WC_Order    $order      Order object.
+	 * @param string|false $permission If the current user can view the order details or not.
+	 * @param array        $attributes Block attributes.
+	 * @param string       $content    Original block content.
+	 *
+	 * @return string
+	 */
+	public function render_content( $order, $permission = false, $attributes = [], $content = '' ) {
+		return parent::render_content( $order, $permission, $attributes, $content );
+	}
+}
