+++ plugins/woocommerce/tests/php/src/Blocks/Mocks/ProductCollectionMock.php
@@ -1,4 +1,6 @@
 <?php
+declare( strict_types = 1 );
+
 namespace Automattic\WooCommerce\Tests\Blocks\Mocks;
 
 use Automattic\WooCommerce\Blocks\BlockTypes\ProductCollection;
