+++ plugins/woocommerce/tests/php/src/Blocks/StoreApi/Utilities/CartControllerTests.php
@@ -1,16 +1,19 @@
 <?php
-/**
- * CartController Tests.
- */
+declare( strict_types = 1 );
 
 namespace Automattic\WooCommerce\Tests\Blocks\StoreApi\Utilities;
 
-use Automattic\WooCommerce\Tests\Blocks\Helpers\FixtureData;
 use Automattic\WooCommerce\StoreApi\Utilities\CartController;
+use Automattic\WooCommerce\Tests\Blocks\Helpers\FixtureData;
 use Yoast\PHPUnitPolyfills\TestCases\TestCase;
 
+/**
+ * Unit tests for the CartController class.
+ */
 class CartControllerTests extends TestCase {
-
+	/**
+	 * Test cart error code is getting exposed.
+	 */
 	public function test_get_cart_errors() {
 		$class    = new CartController();
 		$fixtures = new FixtureData();
@@ -25,7 +28,7 @@ class CartControllerTests extends TestCase {
 		$out_of_stock_product_key = wc()->cart->add_to_cart( $out_of_stock_product->get_id(), 2 );
 		$out_of_stock_in_cart     = wc()->cart->get_cart_item( $out_of_stock_product_key )['data'];
 
-		// This product will have exact levels of stock known
+		// This product will have exact levels of stock known.
 		$partially_out_of_stock_product = $fixtures->get_simple_product(
 			array(
 				'name'          => 'Test Product 2',
@@ -35,7 +38,7 @@ class CartControllerTests extends TestCase {
 		$partially_out_of_stock_key     = wc()->cart->add_to_cart( $partially_out_of_stock_product->get_id(), 4 );
 		$partially_out_of_stock_in_cart = wc()->cart->get_cart_item( $partially_out_of_stock_key )['data'];
 
-		// This product will have exact levels of stock known
+		// This product will have exact levels of stock known.
 		$too_many_in_cart_product     = $fixtures->get_simple_product(
 			array(
 				'name'          => 'Test Product 3',
@@ -50,7 +53,7 @@ class CartControllerTests extends TestCase {
 		$partially_out_of_stock_in_cart->set_stock_quantity( 2 );
 		$too_many_in_cart_in_cart->set_sold_individually( true );
 
-		// This product will not be purchasable
+		// This product will not be purchasable.
 		$not_purchasable_product = $fixtures->get_simple_product(
 			array(
 				'name'          => 'Test Product 4',
@@ -59,10 +62,10 @@ class CartControllerTests extends TestCase {
 		);
 		wc()->cart->add_to_cart( $not_purchasable_product->get_id(), 2 );
 
-		// This function will force the $product->is_purchasable() function to return false for our $not_purchasable_product
+		// This function will force the $product->is_purchasable() function to return false for our $not_purchasable_product.
 		add_filter(
 			'woocommerce_is_purchasable',
-			function( $is_purchasable, $product ) use ( $not_purchasable_product ) {
+			function ( $is_purchasable, $product ) use ( $not_purchasable_product ) {
 				if ( $product->get_id() === $not_purchasable_product->get_id() ) {
 					return false;
 				}
@@ -88,7 +91,5 @@ class CartControllerTests extends TestCase {
 		foreach ( $expected_errors as $expected_error ) {
 			$this->assertContains( $expected_error, $error_codes );
 		}
-
 	}
-
 }
