+++ plugins/woocommerce/tests/php/src/Blocks/StoreApi/Utilities/NoticeHandlerTests.php
@@ -1,7 +1,5 @@
 <?php
-/**
- * NoticeHandler Tests.
- */
+declare( strict_types = 1 );
 
 namespace Automattic\WooCommerce\Tests\Blocks\StoreApi\Utilities;
