+++ plugins/woocommerce/tests/php/src/Internal/Orders/IppFunctionsTest.php
@@ -1,4 +1,5 @@
 <?php
+declare( strict_types = 1 );
 
 namespace Automattic\WooCommerce\Tests\Internal\Orders;
